Added Service Container folder and moved Service Collection Extension to it.
